+++ tests/lib.rs
@@ -517,6 +517,8 @@ fn test_text_object() {
             halign,
             valign,
             text,
+            width,
+            height,
         } => {
             assert_eq!(font_family.as_str(), "sans-serif");
             assert_eq!(*pixel_size, 16);
@@ -538,6 +540,8 @@ fn test_text_object() {
             assert_eq!(*halign, HorizontalAlignment::Center);
             assert_eq!(*valign, VerticalAlignment::Bottom);
             assert_eq!(text.as_str(), "Test");
+            assert_eq!(*width, 87.7188);
+            assert_eq!(*height, 21.7969);
         }
         _ => panic!(),
     };